First solution of scanner
I think it is not the whole scanner, may be we will need some additional token processing for making tokens more accurate and useable. For example we can process id tokens and sort them on key words, function names etc.
Also, this solution doesn't include BVS, which can be also added on this additional token processing.